requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.6.1; python_full_version >= '3.6.2'
boto3==1.24.36
botocore==1.27.36; python_version >= '3.7'
cffi==1.15.1
click==8.1.3; python_version >= '3.7'
cryptography==37.0.4; python_version >= '3.6'
dnspython==2.2.1; python_version >= '3.6' and python_version < '4.0'
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
jwt==1.3.1
mangum==0.15.0
py-bcrypt==0.4
pycparser==2.21
pydantic[email]==1.9.1
pyjwt==2.4.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
s3transfer==0.6.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.19.1; python_version >= '3.6'
typing-extensions==4.3.0; python_version >= '3.7'
urllib3==1.26.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4.0'
uvicorn==0.18.2
